requirements.txt
rembg==2.0.50
Pillow==10.3.0
onnxruntime==1.17.3
numpy==1.26.4
